requirements.txt
streamlit>=1.28.0
openai>=1.3.0
PyMuPDF>=1.23.0
whisper-openai>=20231117
moviepy>=1.0.3
fpdf2>=2.7.6
python-dotenv>=1.0.0
pydub>=0.25.1
streamlit-lottie>=0.0.5
requests>=2.31.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
